viewModel から API に接続する部分を切り離した
